+++ app/src/main/java/com/example/weatherappavito/pages/weather/presentation/WeatherDetailedViewModel.kt
@@ -43,150 +43,150 @@ class WeatherDetailedViewModel(application: Application) : AndroidViewModel(appl
         }
     }
 
-    suspend fun location(context: Context, activity: Activity,int: Int) {
-        fusedLocationProviderClient =
-            LocationServices.getFusedLocationProviderClient(context)
-        requestPermission(context,activity, int)
-        getLastLocation(context,activity, int)
-    }
-
-
-    suspend fun getLastLocation(context: Context, activity: Activity, int: Int) {
-        if (сheckPermission(context)) {
-            if (ActivityCompat.checkSelfPermission(
-                    context,
-                    android.Manifest.permission.ACCESS_COARSE_LOCATION
-                ) == PackageManager.PERMISSION_GRANTED ||
-                ActivityCompat.checkSelfPermission(
-                    context,
-                    android.Manifest.permission.ACCESS_FINE_LOCATION
-                ) == PackageManager.PERMISSION_GRANTED
-            ) {
-                fusedLocationProviderClient =
-                    LocationServices.getFusedLocationProviderClient(context)
-                fusedLocationProviderClient.lastLocation.addOnCompleteListener { task ->
-                    val location: Location? = task.result
-                    if (location == null) {
-                        viewModelScope.launch {
-                            loadData(
-                                newLocationData(
-                                    context,
-                                    activity
-                                )
-                            )
-                        }
-                    } else {
-                        viewModelScope.launch {
-                            loadData(
-                                getCityName(
-                                    location.latitude,
-                                    location.longitude,
-                                    context
-                                )
-                            )
-                        }
-                    }
-                }
-            } else {
-                Toast.makeText(context, "Please Turn on Your device Location", Toast.LENGTH_SHORT)
-                    .show()
-            }
-        } else {
-            viewModelScope.launch {
-                requestPermission(context, activity, int)
-            }
-        }
-    }
-
-
-    //1
-    suspend fun сheckPermission(context: Context): Boolean {
-        if (
-            ActivityCompat.checkSelfPermission(
-                context,
-                android.Manifest.permission.ACCESS_COARSE_LOCATION
-            ) == PackageManager.PERMISSION_GRANTED ||
-            ActivityCompat.checkSelfPermission(
-                context,
-                android.Manifest.permission.ACCESS_FINE_LOCATION
-            ) == PackageManager.PERMISSION_GRANTED
-        ) {
-            return true
-        }
-        return false
-    }
-
-    //2
-
-    suspend fun newLocationData(context: Context, activity: Activity): String {
-
-
-        val locationRequest = LocationRequest()
-        fusedLocationProviderClient =
-            LocationServices.getFusedLocationProviderClient(activity)
-        val a = if (
-            ActivityCompat.checkSelfPermission(
-                context,
-                android.Manifest.permission.ACCESS_COARSE_LOCATION
-            ) == PackageManager.PERMISSION_GRANTED ||
-            ActivityCompat.checkSelfPermission(
-                context,
-                android.Manifest.permission.ACCESS_FINE_LOCATION
-            ) == PackageManager.PERMISSION_GRANTED
-        ) {
-            fusedLocationProviderClient!!.requestLocationUpdates(
-                locationRequest, test(context), Looper.myLooper()
-            ).toString()
-        } else {
-            ""
-        }
-        return a
-    }
-
-    private fun test(context: Context): LocationCallback {
-        val locationCallback = object : LocationCallback() {
-            override fun onLocationResult(locationResult: LocationResult) {
-                val lastLocation: Location = locationResult.lastLocation!!
-//                Log.d(
-//                    "Debug",
-//                    "${
+//    suspend fun location(context: Context, activity: Activity,int: Int) {
+//        fusedLocationProviderClient =
+//            LocationServices.getFusedLocationProviderClient(context)
+//        requestPermission(context,activity, int)
+//        getLastLocation(context,activity, int)
+//    }
+//
+//
+//    suspend fun getLastLocation(context: Context, activity: Activity, int: Int) {
+//        if (сheckPermission(context)) {
+//            if (ActivityCompat.checkSelfPermission(
+//                    context,
+//                    android.Manifest.permission.ACCESS_COARSE_LOCATION
+//                ) == PackageManager.PERMISSION_GRANTED ||
+//                ActivityCompat.checkSelfPermission(
+//                    context,
+//                    android.Manifest.permission.ACCESS_FINE_LOCATION
+//                ) == PackageManager.PERMISSION_GRANTED
+//            ) {
+//                fusedLocationProviderClient =
+//                    LocationServices.getFusedLocationProviderClient(context)
+//                fusedLocationProviderClient.lastLocation.addOnCompleteListener { task ->
+//                    val location: Location? = task.result
+//                    if (location == null) {
 //                        viewModelScope.launch {
-//                            getCityName(
-//                                lastLocation.latitude,
-//                                lastLocation.longitude,
-//                                context
+//                            loadData(
+//                                newLocationData(
+//                                    context,
+//                                    activity
+//                                )
 //                            )
 //                        }
-//                    }"
-//                )
-            }
-        }
-        return locationCallback
-    }
-
-    //3
-    suspend fun getCityName(lat: Double, long: Double, context: Context): String {
-        var cityName: String = ""
-        val geoCoder = Geocoder(context, Locale.getDefault())
-        val Adress = geoCoder.getFromLocation(lat, long, 3)
-
-        cityName = Adress?.get(0)?.locality ?: "Moscow"
-        Log.d("Debug:", cityName)
-        return cityName
-    }
-
-
-    //4
-
-    suspend fun requestPermission(context: Context, activity: Activity, int: Int) {
-        ActivityCompat.requestPermissions(
-            activity,
-            arrayOf(
-                android.Manifest.permission.ACCESS_FINE_LOCATION,
-                android.Manifest.permission.ACCESS_COARSE_LOCATION
-            ),
-            int
-        )
-    }
+//                    } else {
+//                        viewModelScope.launch {
+//                            loadData(
+//                                getCityName(
+//                                    location.latitude,
+//                                    location.longitude,
+//                                    context
+//                                )
+//                            )
+//                        }
+//                    }
+//                }
+//            } else {
+//                Toast.makeText(context, "Please Turn on Your device Location", Toast.LENGTH_SHORT)
+//                    .show()
+//            }
+//        } else {
+//            viewModelScope.launch {
+//                requestPermission(context, activity, int)
+//            }
+//        }
+//    }
+//
+//
+//    //1
+//    suspend fun сheckPermission(context: Context): Boolean {
+//        if (
+//            ActivityCompat.checkSelfPermission(
+//                context,
+//                android.Manifest.permission.ACCESS_COARSE_LOCATION
+//            ) == PackageManager.PERMISSION_GRANTED ||
+//            ActivityCompat.checkSelfPermission(
+//                context,
+//                android.Manifest.permission.ACCESS_FINE_LOCATION
+//            ) == PackageManager.PERMISSION_GRANTED
+//        ) {
+//            return true
+//        }
+//        return false
+//    }
+//
+//    //2
+//
+//    suspend fun newLocationData(context: Context, activity: Activity): String {
+//
+//
+//        val locationRequest = LocationRequest()
+//        fusedLocationProviderClient =
+//            LocationServices.getFusedLocationProviderClient(activity)
+//        val a = if (
+//            ActivityCompat.checkSelfPermission(
+//                context,
+//                android.Manifest.permission.ACCESS_COARSE_LOCATION
+//            ) == PackageManager.PERMISSION_GRANTED ||
+//            ActivityCompat.checkSelfPermission(
+//                context,
+//                android.Manifest.permission.ACCESS_FINE_LOCATION
+//            ) == PackageManager.PERMISSION_GRANTED
+//        ) {
+//            fusedLocationProviderClient!!.requestLocationUpdates(
+//                locationRequest, test(context), Looper.myLooper()
+//            ).toString()
+//        } else {
+//            ""
+//        }
+//        return a
+//    }
+//
+//    private fun test(context: Context): LocationCallback {
+//        val locationCallback = object : LocationCallback() {
+//            override fun onLocationResult(locationResult: LocationResult) {
+//                val lastLocation: Location = locationResult.lastLocation!!
+////                Log.d(
+////                    "Debug",
+////                    "${
+////                        viewModelScope.launch {
+////                            getCityName(
+////                                lastLocation.latitude,
+////                                lastLocation.longitude,
+////                                context
+////                            )
+////                        }
+////                    }"
+////                )
+//            }
+//        }
+//        return locationCallback
+//    }
+//
+//    //3
+//    suspend fun getCityName(lat: Double, long: Double, context: Context): String {
+//        var cityName: String = ""
+//        val geoCoder = Geocoder(context, Locale.getDefault())
+//        val Adress = geoCoder.getFromLocation(lat, long, 3)
+//
+//        cityName = Adress?.get(0)?.locality ?: "Moscow"
+//        Log.d("Debug:", cityName)
+//        return cityName
+//    }
+//
+//
+//    //4
+//
+//    suspend fun requestPermission(context: Context, activity: Activity, int: Int) {
+//        ActivityCompat.requestPermissions(
+//            activity,
+//            arrayOf(
+//                android.Manifest.permission.ACCESS_FINE_LOCATION,
+//                android.Manifest.permission.ACCESS_COARSE_LOCATION
+//            ),
+//            int
+//        )
+//    }
 
 }
